pyproject.toml
[tool.poetry]
name = "lpld"
version = "0.1.0"
description = "Website for the valve company Western Valve."
authors = ["Tibor Leupold <tibor@lpld.io>"]
license = "Proprietary"

[tool.poetry.dependencies]
python = "^3.8"
Django = "~3.2"
wagtail = "~2.16"
whitenoise = "^5.2.0"
python-dotenv = "~0.16.0"
dj-database-url = "~0.5.0"
psycopg2 = "^2.9.1"
gunicorn = "^20.1.0"
django-basic-auth-ip-whitelist = "~0.3.4"
sentry-sdk = "^1.4.3"
django-widget-tweaks = "^1.4.8"
slippers = "^0.3.0"
heroicons = {extras = ["django"], version = "^1.4.0"}
boto3 = "^1.20.41"
django-storages = "^1.12.3"
wagtail-storages = "^0.0.6"
factory-boy = "^3.2.0"
wagtail-factories = "^2.0.1"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
django-debug-toolbar = "^3.2.1"
django-pattern-library = "~0.7.0"
djhtml = "~1.4.10"
flake8 = "^3.8.4"
isort = "^5.6.4"
mypy = "~0.910"
django-stubs = "^1.9.0"
pytest = "^6.2.1"
pytest-django = "^4.1.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "lpld.settings"
addopts = ["--disable-warnings"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
